refactor: 공통 유틸리티 클래스 패키지 이동
- auth 패키지의 AppleClientSecretInjector, JwtTokenProvider, OAuth2SuccessHandler, AppleJwtGenerator 클래스를 common 패키지로 이동
