+++ src/main/kotlin/com/millo/ollim/common/util/AppleJwtGenerator.kt
@@ -1,4 +1,4 @@
-package com.millo.ollim.auth.util
+package com.millo.ollim.common.util
 
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
